+++ src/Shell/Task/HealthcheckTask.php
@@ -245,7 +245,7 @@ class HealthcheckTask extends AppShell
             __('The passbolt config file is present'),
             __('The passbolt config file is missing in {0}', CONFIG),
             [
-                __('Copy {0} to {1}', CONFIG . 'passbolt.php.default', CONFIG 'passbolt.php'),
+                __('Copy {0} to {1}', CONFIG . 'passbolt.php.default', CONFIG . 'passbolt.php'),
                 __('The passbolt config file is not required if passbolt is configured with environment variables')
             ]
         );