+++ CHANGELOG.md
@@ -12,6 +12,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - PASSBOLT-2640: Fixed incomplete urls in email templates
 - PASSBOLT-2640: Improved escaping non safe characters in emails
 - PASSBOLT-2667: Fixed regression: create a user that has been deleted previously returns an error
+- PASSBOLT-2673: Fixed regression: as AD I cannot create a group with the name of previously deleted group
+- PASSBOLT-2545: Fixed regression: As AD deleting a group I should be notified that all members of the group gonna lose access to the passwords shared with the group
 
 ### Added
 - PASSBOLT-2638: Added command to test email configuration and SMTP communication